// warter-pie/src/payment/dto/create-payment.dto.ts
import { ApiProperty } from '@nestjs/swagger';
import { IsInt, IsNotEmpty, IsString, Min } from 'class-validator';

export class CreatePaymentDto {
  @ApiProperty({ description: 'Amount to pay in KRW', example: 10000 })
  @IsInt()
  @Min(100) // Minimum payment amount, adjust as needed
  @IsNotEmpty()
  amount: number;

  @ApiProperty({ description: 'Points to charge', example: 10000 })
  @IsInt()
  @Min(100) // Minimum points to charge, adjust as needed
  @IsNotEmpty()
  chargedPoints: number;

  @ApiProperty({
    description: 'Unique order ID generated by client (or backend)',
    example: 'order_12345',
  })
  @IsString()
  @IsNotEmpty()
  orderId: string; // Client-generated or backend-generated unique ID

  @ApiProperty({
    description: 'Name of the order/product',
    example: '포인트 충전',
  })
  @IsString()
  @IsNotEmpty()
  orderName: string;
}